Fixed bg gradient, white card size, and font size
Font size should respond now to both zoom and screen size.

I changed the SVG width to be max(100%, 1000px) which causes it to scale with screen width, until the width is below 1000px, at which point the SVG stays the same size and starts to get clipped. This prevents the SVGs from shrinking into two tiny worms on small screens. This approach isn't perfect. It looks good on every common mobile device and on desktop, but on tablet aspect ratios, there is a small undesirable gap between the SVGs.