docs: (and refactor) GraphChoice docs and refactor RadialBarChart
BarGraph styleguidist docs now have increment/decrement buttons for series. Importantly, all of the props actually being passed to the graph component get displayed live to the user reading the docs while they change the props with buttons. This is also the case on the GraphChoice component.
I also refactored the handling of props from the RadialBarChart component so that they can take in props of the same exact format as most other graphs, i.e. series is an array of objects each with a 'name' and 'data'.