Mapped colors based on compare table
Added similar mappedColor logic from compareTable component. Added colorIndex and isInCompare variables to verify if the result object is in the compare table as well as using an individual index variable to make sure the colors stay on track based on what is added to the compare table. Added, an index to sortedResults to apply checkboxColor to all checkboxes. Allso added checkboxColor to RowProps and Row.