Created Planner downloading functionality
Imported the html2canvas-pro library, which is used for creating a canvas image based on the DOM for a component, with the pro version allowing html2canvas to work with previously unsupported CSS attributes and properties. Modified the TopMenu component to show a "download" favicon, instead of the "share" favicon, from MUI if isPlanner prop is true. Likewise, changed the tooltip to "Download your Schedule". Then added the onClick functionality for the button, where the html2canvas function is used to convert the referenced component to a canvas image, which is then promptly turned into a blob. This blob is given an objectURL, which is set as the href for a newly-created "a" tag. The "a" tag is added to the DOM, clicked via the script to download the image, then removed from the DOM. The component is referenced through a useRef placed in the PlannerPage component (a new client component wrapping the TopMenu, Planner, and PlannerDownloadUI components). The useRef is passed as a prop to PlannerDownloadUI, which is another new component. The aim of this component is to be able to easily modify the downloaded image, without changing the base page. It contains the PlannerSchedule component, which allows it to be updated with any changes to the planner from the client, as well as any ui design changes. Lastly, the PlannerDownloadUI component is moved very far off screen using "left-[-9999px]", to ensure that it stays in the DOM (hence allowing html2canvas to turn it into a canvas image), while not being visible to the user. The original page.tsx file now only contains the PlannerPage component.