Planner Integration with Compare
Created a separate component for the planner checkbox in @/components/common, as the same code was being reused for both the CompareTable.tsx file and the SearchResultsTable.tsx file. Then added a div wrapping the title of each course with the planner checkbox component so that flexbox could be employed.