Updated UI
Created a separate row at the end for adding to planner functionality and moved the bookmark icon to the row. Created a custom component called PlannerRow, which used the same type as CheckboxRow. However, it did not include two functions: handleClick and removeFromCompare, hence those were made optional, and additional checks were added to ensure that an undefined function would not be called and functionality would remain the same. The PlannerRow component is slightly different from the CheckboxRow component, in that it no longer has a TableSortLabel, as this was redundant. Lastly, the functionality for the button was included in the PlannerRow.